+++ libs/ui/epg/src/index.ts
@@ -5,6 +5,17 @@ export * from './lib/epg-timeline/epg-timeline.component';
 export * from './lib/epg-timeline/epg-timeline-empty-state.component';
 export * from './lib/epg-timeline/epg-timeline-track.component';
 export * from './lib/epg-timeline/epg-timeline.utils';
+// Reusable EPG-view building blocks (shared by the timeline; a future list view
+// can reuse them too).
+export * from './lib/epg-timeline/epg-archive.util';
+export * from './lib/epg-programme-dialog.service';
+export {
+    formatClockTime,
+    summaryHasTimeRange,
+    summaryHasTitle,
+    summaryMinutesLeft,
+    summaryProgress,
+} from './lib/epg-timeline/epg-summary.util';
 export * from './lib/epg-progress-panel/epg-progress-panel.component';
 export * from './lib/epg-source-status/epg-source-status.component';
 export * from './lib/multi-epg/multi-epg-container.component';
